Move navigation between chats and profiles to the `Navigation`.
