+++ client/src/main/kotlin/io/spine/examples/chatspn/desktop/ChatPage.kt
@@ -28,7 +28,6 @@ package io.spine.examples.chatspn.desktop
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.PointerMatcher
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -36,10 +35,8 @@ import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -52,7 +49,6 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.onClick
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.GenericShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.material.Surface
@@ -61,14 +57,8 @@ import androidx.compose.material.icons.filled.Check
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material.icons.filled.Email
-import androidx.compose.material.icons.filled.ExitToApp
-import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.filled.MoreVert
-import androidx.compose.material.icons.filled.Search
 import androidx.compose.material.icons.filled.Send
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.Icon
@@ -92,13 +82,10 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Rect
 import androidx.compose.ui.geometry.RoundRect
 import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Path
 import androidx.compose.ui.graphics.PathOperation
 import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.graphics.drawscope.DrawScope
-import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.input.key.Key
 import androidx.compose.ui.input.key.KeyEventType
@@ -109,7 +96,6 @@ import androidx.compose.ui.input.key.type
 import androidx.compose.ui.input.pointer.PointerButton
 import androidx.compose.ui.input.pointer.PointerIcon
 import androidx.compose.ui.input.pointer.pointerHoverIcon
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
@@ -119,595 +105,57 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Popup
 import com.google.protobuf.Timestamp
 import io.spine.core.UserId
 import io.spine.examples.chatspn.ChatId
 import io.spine.examples.chatspn.MessageId
 import io.spine.examples.chatspn.account.UserProfile
-import io.spine.examples.chatspn.chat.Chat.ChatType
-import io.spine.examples.chatspn.chat.Chat.ChatType.CT_GROUP
 import io.spine.examples.chatspn.chat.Chat.ChatType.CT_PERSONAL
-import io.spine.examples.chatspn.chat.ChatPreview
-import io.spine.examples.chatspn.chat.MessagePreview
 import io.spine.examples.chatspn.message.MessageView
 import io.spine.examples.chatspn.message.event.MessageMarkedAsDeleted
 import java.awt.Cursor
 import java.awt.Cursor.getPredefinedCursor
-import java.text.SimpleDateFormat
 import java.util.*
 import java.util.stream.Collectors
-import kotlin.math.abs
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
 /**
- * Represents the 'Chats' page in the application.
+ * Represents the 'Chat' page in the application.
  *
  * @param client desktop client
  */
 @Composable
 @Preview
-public fun ChatsPage(client: DesktopClient, toRegistration: () -> Unit) {
-    val model = remember { ChatsPageModel(client, toRegistration) }
-    val isProfileTabOpen by remember { model.profileInfoTabState.isVisibleState }
-    Row {
-        LeftSidebar(model)
-        if (isProfileTabOpen) {
-            ProfileTab(model)
-        } else {
-            ChatContent(model)
-        }
-    }
-    ChatDeletionDialog(model)
-    LogoutDialog(model)
-}
-
-/**
- * Represents the left sidebar on the `Chats` page.
- */
-@Composable
-private fun LeftSidebar(model: ChatsPageModel) {
-    Surface(
-        Modifier
-            .padding(end = 1.dp)
-            .drawBehind {
-                drawLine(
-                    color = Color.Gray,
-                    start = Offset(size.width, 0f),
-                    end = Offset(size.width, size.height),
-                    strokeWidth = 1.dp.toPx()
-                )
-            },
-    ) {
-        Column(
-            Modifier
-                .fillMaxHeight()
-                .width(256.dp)
-        ) {
-            Row(
-                Modifier
-                    .fillMaxWidth()
-                    .height(54.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Spacer(Modifier.width(10.dp))
-                MenuButton(model)
-                Spacer(Modifier.width(14.dp))
-                UserSearchField(model)
-            }
-            ChatList(model)
-        }
-    }
-}
-
-/**
- * Represents the menu button.
- */
-@Composable
-private fun MenuButton(model: ChatsPageModel) {
-    val isUserProfileOpen by remember { model.profileInfoTabState.isVisibleState }
-    val isAuthenticatedUser = model.authenticatedUser
-        .equals(model.profileInfoTabState.userProfile.value)
-    val containerColor = if (isUserProfileOpen && isAuthenticatedUser)
-        MaterialTheme.colorScheme.inverseSurface
-    else
-        MaterialTheme.colorScheme.surface
-    Button(
-        modifier = Modifier
-            .size(42.dp),
-        onClick = {
-            if (isUserProfileOpen && isAuthenticatedUser) {
-                model.profileInfoTabState.clear()
-            } else {
-                model.openUserProfileTab(model.authenticatedUser)
-            }
-        },
-        shape = CircleShape,
-        colors = ButtonDefaults.buttonColors(
-            containerColor = containerColor
-        ),
-        contentPadding = PaddingValues(6.dp)
-    ) {
-        Icon(
-            Icons.Default.Menu,
-            "Menu",
-            Modifier.size(20.dp),
-            MaterialTheme.colorScheme.onBackground
+public fun ChatPage(
+    client: DesktopClient,
+    chatState: MutableState<ChatData?>,
+    openModal: (content: @Composable () -> Unit) -> Unit,
+    closeModal: () -> Unit,
+    openChatInfo: (chat: ChatId) -> Unit,
+    openUserProfile: (user: UserId) -> Unit
+) {
+    val model = remember {
+        ChatsPageModel(
+            client,
+            chatState,
+            openModal,
+            closeModal,
+            openChatInfo,
+            openUserProfile
         )
     }
-}
-
-/**
- * Represents the input field to find the user.
- */
-@OptIn(ExperimentalComposeUiApi::class)
-@Composable
-private fun UserSearchField(model: ChatsPageModel) {
-    val viewScope = rememberCoroutineScope { Dispatchers.Default }
-    var inputText by remember { model.userSearchFieldState.userEmailState }
-    var isError by remember { model.userSearchFieldState.errorState }
-    val onSearch = {
-        val email = inputText
-        viewScope.launch {
-            if (email.trim().isNotEmpty()) {
-                model.createPersonalChat(email.trim())
-            }
-        }
-        inputText = ""
-    }
-    BasicTextField(
-        modifier = Modifier
-            .size(182.dp, 30.dp)
-            .background(MaterialTheme.colorScheme.background)
-            .onPreviewKeyEvent {
-                when {
-                    (it.key == Key.Enter) -> {
-                        onSearch()
-                        true
-                    }
-                    else -> false
-                }
-            },
-        value = inputText,
-        singleLine = true,
-        onValueChange = {
-            inputText = it
-            isError = false
-        }
-    ) { innerTextField ->
-        Box(
-            modifier = Modifier
-                .background(
-                    color = MaterialTheme.colorScheme.surface,
-                    shape = RoundedCornerShape(size = 3.dp)
-                )
-                .padding(all = 8.dp)
-        ) {
-            if (inputText.isEmpty()) {
-                Text(
-                    text = "Search",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSecondary
-                )
-            }
-            Row(
-                Modifier.fillMaxWidth(),
-                Arrangement.SpaceBetween,
-                Alignment.CenterVertically
-            ) {
-                Box(Modifier.weight(1f)) {
-                    innerTextField()
-                }
-                if (inputText.trim().isNotEmpty()) {
-                    SearchIcon(onSearch)
-                }
-            }
-        }
-    }
-}
-
-/**
- * Represents the icon button for the `UserSearchField`.
- */
-@Composable
-private fun SearchIcon(onSearch: () -> Unit) {
-    val interactionSource = remember { MutableInteractionSource() }
-    Icon(
-        Icons.Default.Search,
-        "Search",
-        Modifier
-            .padding(horizontal = 4.dp)
-            .pointerHoverIcon(PointerIcon(getPredefinedCursor(Cursor.HAND_CURSOR)))
-            .clickable(
-                interactionSource = interactionSource,
-                indication = null,
-                onClick = onSearch
-            ),
-        MaterialTheme.colorScheme.onSecondary
-    )
-}
-
-/**
- * Represents the list of chat previews.
- */
-@Composable
-private fun ChatList(model: ChatsPageModel) {
-    val chats by model.chats().collectAsState()
-    val selectedChat by model.selectedChat().collectAsState()
-    LazyColumn(
+    model.prepareMessages()
+    Column(
         Modifier.fillMaxSize()
     ) {
-        chats.forEachIndexed { index, chat ->
-            item(key = index) {
-                ChatPreviewPanel(
-                    chat.name,
-                    chat.lastMessage,
-                    chat.id.equals(selectedChat)
-                ) {
-                    model.selectChat(chat.id)
-                }
-            }
+        ChatTopBar(model)
+        Box(Modifier.weight(1f)) {
+            ChatMessages(model)
         }
-    }
-}
-
-/**
- * Represents the chat preview in the chats list.
- */
-@Composable
-private fun ChatPreviewPanel(
-    chatName: String,
-    lastMessage: MessageData?,
-    isSelected: Boolean,
-    select: () -> Unit
-) {
-    Box(
-        Modifier
-            .fillMaxWidth()
-            .clickable { select() }
-            .background(
-                color = if (isSelected)
-                    MaterialTheme.colorScheme.inverseSurface
-                else
-                    MaterialTheme.colorScheme.background
-            ),
-        Alignment.CenterStart,
-    ) {
-        ChatPreviewContent(chatName, lastMessage)
-    }
-}
-
-/**
- * Represents the chat preview content in the chat preview panel.
- */
-@Composable
-private fun ChatPreviewContent(chatName: String, lastMessage: MessageData?) {
-    Row(
-        Modifier.padding(8.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Avatar(46f, chatName)
-        Spacer(Modifier.size(12.dp))
-        Column(horizontalAlignment = Alignment.Start) {
-            Row(
-                Modifier.fillMaxWidth(),
-                Arrangement.SpaceBetween
-            ) {
-                Text(
-                    text = chatName,
-                    color = MaterialTheme.colorScheme.onBackground,
-                    style = MaterialTheme.typography.headlineMedium,
-                )
-                Text(
-                    text = if (lastMessage == null) "" else lastMessage.whenPosted.toStringTime(),
-                    color = MaterialTheme.colorScheme.onSecondary,
-                    style = MaterialTheme.typography.bodySmall,
-                )
-            }
-            Spacer(Modifier.size(9.dp))
-            Text(
-                text = if (lastMessage == null) ""
-                else lastMessage.content.replace("\\s".toRegex(), " "),
-                color = MaterialTheme.colorScheme.onSecondary,
-                style = MaterialTheme.typography.bodyMedium,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis
-            )
-        }
-    }
-}
-
-/**
- * Represents the default avatar.
- */
-@Composable
-public fun Avatar(size: Float, name: String, modifierExtender: Modifier.() -> Modifier = { this }) {
-    val gradients = listOf(
-        listOf(Color(0xFFFFC371), Color(0xFFFF5F6D)),
-        listOf(Color(0xFF95E4FC), Color(0xFF0F65D6)),
-        listOf(Color(0xFF72F877), Color(0xFF259CF1)),
-        listOf(Color(0xFF76DBFA), Color(0xFFD72DFD)),
-        listOf(Color(0xFFA6FA85), Color(0xFF22AC00D)),
-        listOf(Color(0xFFFD9696), Color(0xFFF11010))
-    )
-    val gradientIndex = abs(name.hashCode()) % gradients.size
-    Box(contentAlignment = Alignment.Center) {
-        Image(
-            modifier = Modifier
-                .size(size.dp)
-                .clip(CircleShape)
-                .modifierExtender(),
-            contentScale = ContentScale.Crop,
-            painter = object : Painter() {
-                override val intrinsicSize: Size = Size(size, size)
-                override fun DrawScope.onDraw() {
-                    drawRect(
-                        Brush.linearGradient(gradients[gradientIndex]),
-                        size = Size(size * 4, size * 4)
-                    )
-                }
-            },
-            contentDescription = "User picture"
-        )
-        Text(
-            name[0].toString(),
-            color = Color.White,
-            fontSize = (size * 0.5).sp,
-            style = MaterialTheme.typography.headlineLarge
-        )
-    }
-}
-
-/**
- * Represents the content of the selected chat.
- */
-@Composable
-private fun ChatContent(model: ChatsPageModel) {
-    val selectedChat by model.selectedChat().collectAsState()
-    val chats by model.chats().collectAsState()
-    val isChatSelected = chats
-        .stream()
-        .map { chat -> chat.id }
-        .anyMatch { id -> id.equals(selectedChat) }
-    if (!isChatSelected) {
-        ChatNotChosenBox()
-    } else {
-        Column(
-            Modifier.fillMaxSize()
-        ) {
-            ChatTopBar(model)
-            Box(Modifier.weight(1f)) {
-                ChatMessages(model)
-            }
-            ChatBottomBar(model)
-        }
-    }
-}
-
-/**
- * Represents content when no one chat is selected.
- */
-@Composable
-private fun ChatNotChosenBox() {
-    Box(
-        Modifier.fillMaxSize(),
-        Alignment.Center
-    ) {
-        Text(
-            text = "Choose the chat",
-            style = MaterialTheme.typography.headlineMedium,
-            color = MaterialTheme.colorScheme.onSecondary,
-        )
-    }
-}
-
-/**
- * Represents the info tab with the user profile.
- */
-@Composable
-private fun ProfileTab(model: ChatsPageModel) {
-    Column(
-        Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.surface),
-        Arrangement.Top,
-        Alignment.CenterHorizontally
-    ) {
-        ProfileTopBar(model)
-        ProfileTabContent(model)
-    }
-}
-
-/**
- * Represents the main content of the user profile tab.
- */
-@Composable
-private fun ProfileTabContent(model: ChatsPageModel) {
-    val user by remember { model.profileInfoTabState.userProfile }
-    if (null == user) {
-        model.profileInfoTabState.isVisibleState.value = false
-        return
-    }
-    val chat by remember { model.profileInfoTabState.chatState }
-    val isAuthenticatedUser = model.authenticatedUser
-        .equals(model.profileInfoTabState.userProfile.value)
-    Column(
-        Modifier
-            .widthIn(0.dp, 480.dp)
-            .padding(16.dp, 24.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Avatar(96f, user!!.name)
-        Spacer(Modifier.height(4.dp))
-        Text(
-            user!!.name,
-            color = MaterialTheme.colorScheme.onBackground,
-            style = MaterialTheme.typography.headlineLarge
-        )
-        Spacer(Modifier.height(12.dp))
-        if (!isAuthenticatedUser) {
-            ProfileTabMessageButton(model)
-            Spacer(Modifier.height(8.dp))
-        }
-        ProfileTabEmailField(user!!.email.value)
-        Spacer(Modifier.height(8.dp))
-        if (isAuthenticatedUser) {
-            ProfileTabLogOutButton(model)
-        } else if (null != chat) {
-            ProfileTabDeleteChatButton(model)
-        }
-    }
-}
-
-/**
- * Represents a button for opening a chat with a user on the user's profile tab.
- */
-@Composable
-private fun ProfileTabMessageButton(model: ChatsPageModel) {
-    val user by remember { model.profileInfoTabState.userProfile }
-    val chat by remember { model.profileInfoTabState.chatState }
-    InfoTabButton(
-        "Message",
-        Icons.Default.Send
-    ) {
-        if (null != chat) {
-            model.selectChat(chat!!.id)
-        } else {
-            model.createPersonalChat(user!!.id)
-        }
-    }
-}
-
-/**
- * Represents a button for logging out on the authenticated user profile tab.
- */
-@Composable
-private fun ProfileTabLogOutButton(model: ChatsPageModel) {
-    InfoTabButton(
-        "Log out",
-        Icons.Default.ExitToApp,
-        MaterialTheme.colorScheme.error
-    ) {
-        model.isLogoutDialogVisible.value = true
-    }
-}
-
-/**
- * Represents a button to delete a chat with a user on the user profile tab.
- */
-@Composable
-private fun ProfileTabDeleteChatButton(model: ChatsPageModel) {
-    val chat by remember { model.profileInfoTabState.chatState }
-    InfoTabButton(
-        "Delete chat",
-        Icons.Default.Delete,
-        MaterialTheme.colorScheme.error
-    ) {
-        model.chatInDeletionState.value = chat
-    }
-}
-
-/**
- * Represents an email field on the user profile tab.
- */
-@Composable
-private fun ProfileTabEmailField(email: String) {
-    Box(Modifier.clip(MaterialTheme.shapes.small)) {
-        Column(
-            Modifier
-                .background(MaterialTheme.colorScheme.background)
-                .fillMaxWidth()
-                .padding(12.dp, 8.dp),
-            verticalArrangement = Arrangement.Center
-        ) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Icon(
-                    Icons.Default.Email,
-                    "Email",
-                    Modifier.size(20.dp),
-                    MaterialTheme.colorScheme.primary
-                )
-                Spacer(Modifier.width(4.dp))
-                Text(
-                    "Email",
-                    color = MaterialTheme.colorScheme.primary,
-                    style = MaterialTheme.typography.headlineMedium
-                )
-            }
-            Spacer(Modifier.height(4.dp))
-            Text(
-                email,
-                style = MaterialTheme.typography.bodyLarge
-            )
-        }
-
-    }
-}
-
-/**
- * Represents a button on an info tab.
- */
-@Composable
-private fun InfoTabButton(
-    text: String = "",
-    icon: ImageVector? = null,
-    contentColor: Color = MaterialTheme.colorScheme.primary,
-    onClick: () -> Unit
-) {
-    Box(Modifier.clip(MaterialTheme.shapes.small)) {
-        Button(
-            onClick,
-            Modifier
-                .fillMaxWidth()
-                .pointerHoverIcon(PointerIcon(getPredefinedCursor(Cursor.HAND_CURSOR))),
-            shape = MaterialTheme.shapes.small,
-            colors = ButtonDefaults.buttonColors(
-                containerColor = MaterialTheme.colorScheme.background,
-                contentColor = contentColor
-            ),
-            contentPadding = PaddingValues(12.dp, 8.dp)
-        ) {
-            Row(Modifier.fillMaxWidth(), verticalAlignment = Alignment.CenterVertically) {
-                if (icon != null) {
-                    Icon(
-                        icon,
-                        text,
-                        Modifier.size(20.dp),
-                        contentColor
-                    )
-                    Spacer(Modifier.width(4.dp))
-                }
-                Text(
-                    text,
-                    color = contentColor,
-                    style = MaterialTheme.typography.bodyLarge
-                )
-            }
-        }
-    }
-}
-
-/**
- * Represents the top bar on the profile tab.
- */
-@Composable
-private fun ProfileTopBar(model: ChatsPageModel) {
-    TopBar {
-        Row(
-            Modifier.fillMaxWidth().padding(horizontal = 6.dp),
-            Arrangement.SpaceBetween,
-            Alignment.CenterVertically
-        ) {
-            TextButton("< Back") {
-                model.profileInfoTabState.clear()
-            }
-            Text("Info", style = MaterialTheme.typography.headlineMedium)
-            Spacer(Modifier.width(76.dp))
-        }
+        ChatBottomBar(model)
     }
 }
 
@@ -716,8 +164,7 @@ private fun ProfileTopBar(model: ChatsPageModel) {
  */
 @Composable
 private fun ChatTopBar(model: ChatsPageModel) {
-    val selectedChat by model.selectedChat().collectAsState()
-    val chat = model.getChatData(selectedChat)!!
+    val chatData = remember { model.chatState }
     val interactionSource = remember { MutableInteractionSource() }
     TopBar {
         Row(
@@ -732,42 +179,19 @@ private fun ChatTopBar(model: ChatsPageModel) {
                         interactionSource = interactionSource,
                         indication = null
                     ) {
-                        model.openChatInfoTab(chat.id)
+                        model.openChatInfo(chatData.value!!.id)
                     },
                 verticalAlignment = Alignment.CenterVertically
             ) {
-                Avatar(42f, chat.name)
+                Avatar(42f, chatData.value!!.name)
                 Text(
-                    chat.name,
+                    chatData.value!!.name,
                     modifier = Modifier.padding(start = 8.dp),
                     style = MaterialTheme.typography.headlineMedium,
                 )
             }
-            ChatMoreButton(model, chat)
+            ChatMoreButton(model, chatData.value!!)
         }
-    }
-}
-
-/**
- * Represents the top bar with the configurable content.
- */
-@Composable
-private fun TopBar(content: @Composable () -> Unit) {
-    Surface(
-        Modifier
-            .fillMaxWidth()
-            .height(54.dp)
-            .padding(bottom = 1.dp)
-            .drawBehind {
-                drawLine(
-                    color = Color.Gray,
-                    start = Offset(0f, size.height),
-                    end = Offset(size.width, size.height),
-                    strokeWidth = 1.dp.toPx()
-                )
-            },
-    ) {
-        content()
     }
 }
 
@@ -811,24 +235,32 @@ private fun ChatDropdownMenu(
             Icons.Default.Delete,
             MaterialTheme.colorScheme.error
         ) {
-            model.chatInDeletionState.value = chat
+            model.openModal(
+                ChatDeletionDialog(
+                    {
+                        model.closeModal()
+                    },
+                    {
+                        model.deleteChat(chat.id)
+                    },
+                    chat
+                )
+            )
             isMenuOpen.value = false
         }
     }
 }
 
 /**
- * Represents a chat deletion confirmation modal.
+ * Returns a representation of the chat deletion confirmation modal.
  */
-@Composable
-private fun ChatDeletionDialog(
-    model: ChatsPageModel,
-) {
-    val viewScope = rememberCoroutineScope { Dispatchers.Default }
-    val chat by remember { model.chatInDeletionState }
-    val isVisible = remember { mutableStateOf(false) }
-    isVisible.value = model.chatInDeletionState.value != null
-    ModalWindow(isVisible, { model.chatInDeletionState.value = null }) {
+public fun ChatDeletionDialog(
+    onDismiss: () -> Unit,
+    onDelete: () -> Unit,
+    chat: ChatData
+): @Composable () -> Unit {
+    return {
+        val viewScope = rememberCoroutineScope()
         Column(
             Modifier.width(300.dp)
                 .padding(start = 16.dp, top = 24.dp, end = 16.dp, bottom = 8.dp)
@@ -836,12 +268,12 @@ private fun ChatDeletionDialog(
             Text(
                 buildAnnotatedString {
                     append("Are you sure you want to delete chat ")
-                    if (chat?.type == CT_PERSONAL) {
+                    if (chat.type == CT_PERSONAL) {
                         append("with ")
                     }
                     append(
                         AnnotatedString(
-                            chat?.name ?: "",
+                            chat.name,
                             spanStyle = SpanStyle(fontWeight = FontWeight.Bold)
                         )
                     )
@@ -860,95 +292,16 @@ private fun ChatDeletionDialog(
                 Arrangement.spacedBy(8.dp, Alignment.End)
             ) {
                 TextButton("Cancel") {
-                    model.chatInDeletionState.value = null
+                    onDismiss()
                 }
                 TextButton("Delete", MaterialTheme.colorScheme.error) {
-                    val id = chat!!.id
                     viewScope.launch {
-                        model.deleteChat(id)
+                        onDelete()
                     }
-                    model.chatInDeletionState.value = null
+                    onDismiss()
                 }
             }
         }
-    }
-}
-
-/**
- * Represents a logout confirmation modal.
- */
-@Composable
-private fun LogoutDialog(
-    model: ChatsPageModel,
-) {
-    val viewScope = rememberCoroutineScope { Dispatchers.Default }
-    val isVisible = remember { model.isLogoutDialogVisible }
-    ModalWindow(isVisible, { isVisible.value = false }) {
-        Column(
-            Modifier.width(300.dp)
-                .padding(start = 16.dp, top = 24.dp, end = 16.dp, bottom = 8.dp)
-        ) {
-            Text(
-                "Are you sure you want to log out?",
-                style = MaterialTheme.typography.bodyLarge
-            )
-            Spacer(Modifier.height(8.dp))
-            Row(
-                Modifier.fillMaxWidth(),
-                Arrangement.spacedBy(8.dp, Alignment.End)
-            ) {
-                TextButton("Cancel") {
-                    isVisible.value = false
-                }
-                TextButton("Log out", MaterialTheme.colorScheme.error) {
-                    viewScope.launch {
-                        model.logOut()
-                    }
-                    isVisible.value = false
-                }
-            }
-        }
-    }
-}
-
-/**
- * Represents the text button without a background and with an optional icon.
- *
- * @param text text on the button
- * @param contentColor color of the text and icon
- * @param icon icon to be displayed before text
- * @param onClick callback that will be triggered when the button clicked
- */
-@Composable
-private fun TextButton(
-    text: String,
-    contentColor: Color = MaterialTheme.colorScheme.primary,
-    icon: ImageVector? = null,
-    onClick: () -> Unit
-) {
-    Button(
-        onClick,
-        Modifier.pointerHoverIcon(PointerIcon(getPredefinedCursor(Cursor.HAND_CURSOR))),
-        shape = MaterialTheme.shapes.small,
-        colors = ButtonDefaults.buttonColors(
-            containerColor = MaterialTheme.colorScheme.background,
-            contentColor = contentColor
-        ),
-        contentPadding = PaddingValues(16.dp, 4.dp)
-    ) {
-        if (icon != null) {
-            Icon(
-                icon,
-                text,
-                Modifier.size(20.dp),
-                contentColor
-            )
-            Spacer(Modifier.width(4.dp))
-        }
-        Text(
-            text,
-            style = MaterialTheme.typography.bodyLarge
-        )
     }
 }
 
@@ -957,9 +310,9 @@ private fun TextButton(
  */
 @Composable
 private fun ChatMessages(model: ChatsPageModel) {
-    val selectedChat by model.selectedChat().collectAsState()
+    val chat = remember { model.chatState }
     val messages by model
-        .messages(selectedChat)
+        .messages()
         .collectAsState()
     LazyColumn(
         Modifier
@@ -1170,7 +523,7 @@ private fun MessageSenderAvatar(model: ChatsPageModel, isVisible: Boolean, user:
                         interactionSource = interactionSource,
                         indication = null
                     ) {
-                        model.openUserProfileTab(user)
+                        model.openUserProfile(user.id)
                     }
             }
             Spacer(Modifier.width(4.dp))
@@ -1299,15 +652,6 @@ private fun PostedTime(time: Timestamp) {
         style = MaterialTheme.typography.bodySmall,
         color = MaterialTheme.colorScheme.onSecondary
     )
-}
-
-/**
- * Converts `Timestamp` to the `hh:mm` string.
- */
-private fun Timestamp.toStringTime(): String {
-    val date = Date(this.seconds * 1000)
-    val format = SimpleDateFormat("hh:mm", Locale.getDefault())
-    return format.format(date)
 }
 
 /**
@@ -1496,128 +840,40 @@ private fun EditMessagePanel(model: ChatsPageModel) {
 }
 
 /**
- * Represents the default modal window.
- *
- * @param isVisibleState mutable state of modal window visibility
- * @param content content of the modal window
- */
-@Composable
-private fun ModalWindow(
-    isVisibleState: MutableState<Boolean>,
-    onDismiss: () -> Unit = {},
-    content: @Composable () -> Unit
-) {
-    if (isVisibleState.value) {
-        ShadedBackground()
-        Popup(
-            alignment = Alignment.Center,
-            onDismissRequest = {
-                isVisibleState.value = false
-                onDismiss()
-            },
-            focusable = true
-        ) {
-            Surface(
-                elevation = 8.dp,
-                shape = MaterialTheme.shapes.small
-            ) {
-                content()
-            }
-        }
-    }
-}
-
-/**
- * Represents the partially transparent black background.
- *
- * The user cannot click on elements behind it.
- */
-@Composable
-private fun ShadedBackground() {
-    Popup {
-        Box(
-            Modifier
-                .fillMaxSize()
-                .background(Color(0x33000000))
-        )
-    }
-}
-
-/**
- * UI Model for the [ChatsPage].
+ * UI Model for the [ChatPage].
  *
  * UI Model is a layer between `@Composable` functions and client.
  */
 private class ChatsPageModel(
     private val client: DesktopClient,
-    private val toRegistration: () -> Unit
+    val chatState: MutableState<ChatData?>,
+    val openModal: (content: @Composable () -> Unit) -> Unit,
+    val closeModal: () -> Unit,
+    val openChatInfo: (chat: ChatId) -> Unit,
+    val openUserProfile: (user: UserId) -> Unit
 ) {
-    private val selectedChatState = MutableStateFlow<ChatId>(ChatId.getDefaultInstance())
-    private val chatPreviewsState = MutableStateFlow<ChatList>(listOf())
-    private val chatMessagesStateMap: MutableMap<ChatId, MutableMessagesState> = mutableMapOf()
-    val userSearchFieldState: UserSearchFieldState = UserSearchFieldState()
+    private val messagesState: MutableMessagesState = MutableStateFlow(listOf())
     val messageInputFieldState: MessageInputFieldState = MessageInputFieldState()
-    val chatInDeletionState: MutableState<ChatData?> = mutableStateOf(null)
-    val isLogoutDialogVisible: MutableState<Boolean> = mutableStateOf(false)
-    val profileInfoTabState: ProfileInfoTabState = ProfileInfoTabState()
-    val authenticatedUser: UserProfile = client.authenticatedUser!!
+    val authenticatedUser: UserProfile
+        get() {
+            return client.authenticatedUser!!
+        }
 
-    init {
-        updateChats(client.readChats().toChatDataList(client))
-        client.observeChats { state -> updateChats(state.chatList.toChatDataList(client)) }
-    }
-
-    /**
-     * Logs out the user and navigates to the registration.
-     */
-    fun logOut() {
-        client.logOut()
-        toRegistration()
-    }
-
-    /**
-     * Returns the state of the user's chats.
-     */
-    fun chats(): StateFlow<ChatList> {
-        return chatPreviewsState
+    fun prepareMessages() {
+        messageInputFieldState.clear()
+        updateMessages(client.readMessages(chatState.value!!.id).toMessageDataList(client))
+        client.stopObservingMessages()
+        client.observeMessages(
+            chatState.value!!.id,
+            { messageView -> updateMessagesState(messageView) },
+            { messageDeleted -> updateMessagesState(messageDeleted) })
     }
 
     /**
      * Returns the state of messages in the chat.
-     *
-     * @param chat ID of the chat to get messages from
      */
-    fun messages(chat: ChatId): MessagesState {
-        if (!chatMessagesStateMap.containsKey(chat)) {
-            throw IllegalStateException("Chat not found")
-        }
-        return chatMessagesStateMap[chat]!!
-    }
-
-    /**
-     * Returns the state of the selected chat.
-     */
-    fun selectedChat(): StateFlow<ChatId> {
-        return selectedChatState
-    }
-
-    /**
-     * Selects provided chat and subscribes to message changes in it.
-     *
-     * Also clears the message input field state.
-     *
-     * @param chat ID of the chat to select
-     */
-    fun selectChat(chat: ChatId) {
-        selectedChatState.value = chat
-        messageInputFieldState.clear()
-        profileInfoTabState.clear()
-        updateMessages(chat, client.readMessages(chat).toMessageDataList(client))
-        client.stopObservingMessages()
-        client.observeMessages(
-            chat,
-            { messageView -> chat.updateMessagesState(messageView) },
-            { messageDeleted -> chat.updateMessagesState(messageDeleted) })
+    fun messages(): MessagesState {
+        return messagesState
     }
 
     /**
@@ -1626,14 +882,14 @@ private class ChatsPageModel(
      *
      * @param messageView message to update the state
      */
-    private fun ChatId.updateMessagesState(messageView: MessageView) {
+    private fun updateMessagesState(messageView: MessageView) {
         val message = messageView.toMessageData(client)
-        val chatMessages = chatMessagesStateMap[this]!!.value
+        val chatMessages = messagesState.value
         if (chatMessages.findMessage(message.id) != null) {
             val newChatMessages = chatMessages.replaceMessage(message)
-            updateMessages(this, newChatMessages)
+            updateMessages(newChatMessages)
         } else {
-            updateMessages(this, chatMessages + message)
+            updateMessages(chatMessages + message)
         }
     }
 
@@ -1642,13 +898,13 @@ private class ChatsPageModel(
      *
      * @param messageDeleted event about message deletion
      */
-    private fun ChatId.updateMessagesState(messageDeleted: MessageMarkedAsDeleted) {
-        val chatMessages = chatMessagesStateMap[this]!!.value
+    private fun updateMessagesState(messageDeleted: MessageMarkedAsDeleted) {
+        val chatMessages = messagesState.value
         val message = chatMessages.findMessage(messageDeleted.id)
         if (message != null) {
             val messageIndex = chatMessages.indexOf(message)
             val newChatMessages = chatMessages.remove(messageIndex)
-            updateMessages(this, newChatMessages)
+            updateMessages(newChatMessages)
         }
     }
 
@@ -1678,19 +934,7 @@ private class ChatsPageModel(
      * @param content message text content
      */
     fun sendMessage(content: String) {
-        client.sendMessage(selectedChatState.value, content)
-    }
-
-    /**
-     * Deletes the chat.
-     *
-     * @param chat ID of the chat to delete
-     */
-    fun deleteChat(chat: ChatId) {
-        client.deleteChat(chat)
-        if (profileInfoTabState.chatState.value?.id?.equals(chat) ?: false) {
-            profileInfoTabState.chatState.value = null
-        }
+        client.sendMessage(chatState.value!!.id, content)
     }
 
     /**
@@ -1699,7 +943,7 @@ private class ChatsPageModel(
      * @param message ID of the message to remove
      */
     fun removeMessage(message: MessageId) {
-        client.removeMessage(selectedChatState.value, message)
+        client.removeMessage(chatState.value!!.id, message)
     }
 
     /**
@@ -1709,117 +953,25 @@ private class ChatsPageModel(
      * @param newContent new text content for the message
      */
     fun editMessage(message: MessageId, newContent: String) {
-        client.editMessage(selectedChatState.value, message, newContent)
-    }
-
-    /**
-     * Creates the personal chat between the authenticated and the provided user.
-     *
-     * Selects the created chat.
-     *
-     * @param user ID of the user with whom to create a personal chat
-     */
-    fun createPersonalChat(user: UserId) {
-        client.createPersonalChat(user) { event ->
-            selectChat(event.id)
-        }
-    }
-
-    /**
-     * Finds user by email and creates the personal chat between authenticated and found user.
-     *
-     * @param email email of the user with whom to create a personal chat
-     */
-    fun createPersonalChat(email: String) {
-        val user = client.findUser(email)
-        if (null != user) {
-            client.createPersonalChat(user.id)
-        } else {
-            userSearchFieldState.errorState.value = true
-        }
-    }
-
-    /**
-     * Opens a tab with a user profile.
-     *
-     * @param user profile of the user to open
-     */
-    fun openUserProfileTab(user: UserProfile) {
-        profileInfoTabState.isVisibleState.value = true
-        profileInfoTabState.userProfile.value = user
-        profileInfoTabState.chatState.value = findPersonalChat(user.id)
-    }
-
-    /**
-     * Opens a tab with chat info. If the chat is personal, a user profile will be opened.
-     *
-     * @param chatId ID of the chat which info to open
-     */
-    fun openChatInfoTab(chatId: ChatId) {
-        val chat = chatPreviewsState.value.find { chatData -> chatId.equals(chatData.id) } ?: return
-        if (chat.type == CT_PERSONAL) {
-            val userId = chat.members.find { user -> !user.equals(authenticatedUser.id) }
-            val user = client.findUser(userId!!)
-            profileInfoTabState.isVisibleState.value = true
-            profileInfoTabState.userProfile.value = user
-            profileInfoTabState.chatState.value = getChatData(chat.id)
-        }
-    }
-
-    /**
-     * Returns the data of the personal chat with the provided user,
-     * or `null` if the chat doesn't exist.
-     *
-     * @param user ID of the user with whom to find the personal chat
-     */
-    private fun findPersonalChat(user: UserId): ChatData? {
-        val chat = chatPreviewsState.value.find { chatData ->
-            chatData.type == CT_PERSONAL && chatData.members.contains(user)
-        }
-        return chat
-    }
-
-    /**
-     * Returns the data of the chat by id, or `null` if the chat doesn't exist.
-     *
-     * @param chatId ID of the chat
-     */
-    fun getChatData(chatId: ChatId): ChatData? {
-        val chat = chatPreviewsState.value.find { chatData ->
-            chatData.id.equals(chatId)
-        }
-        return chat
-    }
-
-    /**
-     * Updates the model with new chats.
-     *
-     * @param chats new list of user chats
-     */
-    private fun updateChats(chats: ChatList) {
-        chatPreviewsState.value = chats
+        client.editMessage(chatState.value!!.id, message, newContent)
     }
 
     /**
      * Updates the model with new messages.
      *
-     * @param chat ID of the chat to update messages in
      * @param messages new list of messages in the chat
      */
-    private fun updateMessages(chat: ChatId, messages: MessageList) {
-        if (chatMessagesStateMap.containsKey(chat)) {
-            chatMessagesStateMap[chat]!!.value = messages
-        } else {
-            chatMessagesStateMap[chat] = MutableStateFlow(messages)
-        }
+    private fun updateMessages(messages: MessageList) {
+        messagesState.value = messages
     }
 
     /**
-     * State of the user search field.
+     * Deletes the chat.
+     *
+     * @param chat ID of the chat to delete
      */
-    class UserSearchFieldState {
-        val userEmailState: MutableState<String> = mutableStateOf("")
-        val errorState: MutableState<Boolean> = mutableStateOf(false)
+    fun deleteChat(chat: ChatId) {
+        client.deleteChat(chat)
     }
 
     /**
@@ -1839,36 +991,7 @@ private class ChatsPageModel(
             editingMessage.value = null
         }
     }
-
-    /**
-     * State of the modal window with the user profile.
-     */
-    class ProfileInfoTabState {
-        val userProfile: MutableState<UserProfile?> = mutableStateOf(null)
-        val chatState: MutableState<ChatData?> = mutableStateOf(null)
-        val isVisibleState: MutableState<Boolean> = mutableStateOf(false)
-
-        /**
-         * Clears the state.
-         */
-        fun clear() {
-            userProfile.value = null
-            chatState.value = null
-            isVisibleState.value = false
-        }
-    }
 }
-
-/**
- * Data for the chat preview.
- */
-public data class ChatData(
-    val id: ChatId,
-    val name: String,
-    val lastMessage: MessageData?,
-    val members: List<UserId>,
-    val type: ChatType
-)
 
 /**
  * Data for the single message view.
@@ -1911,67 +1034,6 @@ private fun MessageList.replaceMessage(newMessage: MessageData): MessageList {
 }
 
 /**
- * Creates the `ChatData` list from the `ChatPreview` list.
- *
- * @param client desktop client to find user profiles
- */
-private fun List<ChatPreview>.toChatDataList(client: DesktopClient): ChatList {
-    return this.stream().map { chatPreview ->
-        ChatData(
-            chatPreview.id,
-            chatPreview.name(client),
-            chatPreview.lastMessageData(client),
-            chatPreview.members(client),
-            chatPreview.type()
-        )
-    }.collect(Collectors.toList())
-}
-
-/**
- * Retrieves the last message from the `ChatPreview` and creates the `MessageData` from it.
- */
-private fun ChatPreview.lastMessageData(client: DesktopClient): MessageData? {
-    val isMessageDefault = this.lastMessage.equals(MessagePreview.getDefaultInstance())
-    return if (isMessageDefault) {
-        null
-    } else {
-        this.lastMessage.toMessageData(client)
-    }
-}
-
-/**
- * Retrieves chat type.
- */
-private fun ChatPreview.type(): ChatType {
-    return if (this.hasGroupChat()) {
-        CT_GROUP
-    } else {
-        CT_PERSONAL
-    }
-}
-
-/**
- * Retrieves members of the chat.
- */
-private fun ChatPreview.members(client: DesktopClient): List<UserId> {
-    return client.readChatMembers(this.id)
-}
-
-/**
- * Creates the `MessageData` from the `MessagePreview`.
- *
- * @param client desktop client to find user profiles
- */
-private fun MessagePreview.toMessageData(client: DesktopClient): MessageData {
-    return MessageData(
-        this.id,
-        client.findUser(this.user)!!,
-        this.content,
-        this.whenPosted
-    )
-}
-
-/**
  * Creates the `MessageData` from the `MessageView`.
  *
  * @param client desktop client to find user profiles
@@ -1983,23 +1045,6 @@ private fun MessageView.toMessageData(client: DesktopClient): MessageData {
         this.content,
         this.whenPosted
     )
-}
-
-/**
- * Retrieves the display name of the chat.
- *
- * @param client desktop client to find user profiles
- */
-private fun ChatPreview.name(client: DesktopClient): String {
-    if (this.hasGroupChat()) {
-        return this.groupChat.name
-    }
-    val creator = this.personalChat.creator
-    val member = this.personalChat.member
-    if (client.authenticatedUser!!.id.equals(creator)) {
-        return client.findUser(member)!!.name
-    }
-    return client.findUser(creator)!!.name
 }
 
 /**
@@ -2026,11 +1071,6 @@ private fun List<MessageView>.toMessageDataList(client: DesktopClient): MessageL
     }.collect(Collectors.toList())
     return messages
 }
-
-/**
- * List of `ChatData`.
- */
-public typealias ChatList = List<ChatData>
 
 /**
  * List of `MessageData`.
